+++ samples/detect_samp_freq/main.cpp
@@ -66,7 +66,7 @@ int main()
         .pio_sm = 0,
         .dma_channel0 = 0,
         .dma_channel1 = 1,
-        .alarm = 0,
+        .alarm_pool = alarm_pool_get_default(),
         .flags = SPDIF_RX_FLAGS_ALL
     };